refactor: extract GPU probe logic into gpu_selector module
Rollback trigger: issues with GPU free space estimation or NVIDIA SMI interactions.
